Refactoring UI libs - Standartize Window scale behaviour across platforms - CursorType - take enties from Avalonia

+++ dotNET/PdfClown.UI.Core/TouchEventArgs.cs
@@ -14,7 +14,7 @@ namespace PdfClown.UI
 
         public TouchAction ActionType { get; }
         public MouseButton MouseButton { get; }
-        public int WheelDelta { get; set; }
+        public float WheelDelta { get; set; }
         public SKPoint Location { get; set; }
         public long PointerId { get; set; }
         public KeyModifiers KeyModifiers { get; set; }
